requirements.txt
#执行以下命令安装依赖库：pip3  install -r requirements.txt  
#查看具体库的版本： pip show httpx

llama-index
# 处理docx文件所需，暂时无用
docx2txt
python-dotenv
#openai 库暂时没有用到
openai
# 使用ollama的模型
llama-index-llms-ollama
# 使用huggingface中的embedding模型
llama-index-embeddings-huggingface
# 添加 llama-index 的 Chroma 向量存储支持
llama-index-vector-stores-chroma
# 添加 llama-index 的 Milvus 向量存储支持
llama-index-vector-stores-milvus

# ZhipuAI API 客户端
zhipuai==2.1.5.20230904

# 使用socks代理
httpx[socks]

chromadb==0.5.12

pymilvus==2.4.7

flask==3.0.3

transformers==4.41.2

#spacy==3.8.2
